+++ classes/steps/GradesCleanup.php
@@ -30,7 +30,6 @@ use moodle_database;
  * @license    https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class GradesCleanup extends AbstractCleanupStep {
-
     /**
      * Default number of days to keep grade history records.
      */